requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
django-cors-headers
django-filter